transform: Send Slack message with future dates
Sometimes we received data from GISAID with collection dates in the
future. This causes a big headache for the Nextstrain nCoV build
maintainers when left uncaught until the end of a multi hour long build.

Send a Slack message containing the GISAID EPI ISL ID and collection
date for every strain that has a collection date from the future (e.g.
a date > now). Emma specifically requested that this Slack message be
formatted differently from the others so that it stands out.