Generate an exclude.txt file
Create an automatically generated `exclude.txt` file in the style of the
Nextstrain, manually curated, `exclude.txt`.

For now, exclude only strains with collection dates set in the future.
Continue to notify Slack about these dropped sequences, but use the
existing `notify-slack` script to do so.

Emma asked for a differently formatted Slack message that would make
exclusions stand out, so send an additional Slack message with an emoji
and some text alerting the user to the new exclusions. I could have
edited the `notify-slack` script to accept an optional text argument
when `--upload` is used, but it seemed easier to just send two Slack
messages instead.